refactor: Update test packages.
