+++ src/test/java/com/limechain/consensus/babe/scale/message/BabeConsensusMessageReaderTest.java
@@ -1,9 +1,8 @@
-package com.limechain.babe.scale;
+package com.limechain.consensus.babe.scale.message;
 
 import com.limechain.chain.lightsyncstate.BabeEpoch;
 import com.limechain.consensus.babe.dto.message.BabeConsensusMessage;
 import com.limechain.consensus.babe.dto.message.BabeConsensusMessageFormat;
-import com.limechain.consensus.babe.scale.message.BabeConsensusMessageReader;
 import com.limechain.utils.StringUtils;
 import com.limechain.utils.scale.ScaleUtils;
 import org.junit.jupiter.api.Test;
