+++ src/test/java/com/limechain/consensus/babe/scale/runtime/BabeApiConfigurationReaderTest.java
@@ -1,8 +1,7 @@
-package com.limechain.babe.scale;
+package com.limechain.consensus.babe.scale.runtime;
 
 import com.limechain.chain.lightsyncstate.BabeEpoch;
 import com.limechain.consensus.babe.dto.runtime.BabeApiConfiguration;
-import com.limechain.consensus.babe.scale.runtime.BabeApiConfigurationReader;
 import com.limechain.utils.StringUtils;
 import com.limechain.utils.scale.ScaleUtils;
 import org.junit.jupiter.api.Test;
