+++ src/test/java/com/limechain/consensus/grandpa/GrandpaSetStateTest.java
@@ -1,7 +1,6 @@
-package com.limechain.grandpa.state;
+package com.limechain.consensus.grandpa;
 
 import com.limechain.consensus.dto.Authority;
-import com.limechain.consensus.grandpa.GrandpaSetState;
 import com.limechain.storage.KVRepository;
 import com.limechain.storage.block.state.BlockState;
 import com.limechain.storage.crypto.KeyStore;
